Changed the name from blogpost to recipe and then updated server.js. Tried running but middleware errors.
