+++ server/package.json
@@ -10,6 +10,7 @@
   "author": "Dean, John, Mac",
   "license": "ISC",
   "dependencies": {
+    "cors": "^2.8.5",
     "express": "^4.17.1",
     "mongoose": "^5.9.4"
   }
